Restructured the project and added Makefile.
