+++ bugzooka/analysis/__init__.py
@@ -1,0 +1,1 @@
+"""Log analysis and processing functionality."""
