+++ bugzooka/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core application functionality and configuration."""
