+++ bugzooka/core/config.py
@@ -1,9 +1,10 @@
 import json
 import os
+import logging.config
 from dotenv import load_dotenv
 
-from src.prompts import GENERIC_APP_PROMPT
-from src.constants import (
+from bugzooka.analysis.prompts import GENERIC_APP_PROMPT
+from bugzooka.core.constants import (
     GENERIC_INFERENCE_URL,
     GENERIC_MODEL,
     INFERENCE_API_RETRY_ATTEMPTS,
@@ -89,3 +90,41 @@ def get_product_config(product_name: str):
             ),
         },
     }
+
+
+def configure_logging(log_level):
+    """
+    Configure application logging.
+
+    param log_level: log level for logging
+    return: None
+    """
+    log_msg_fmt = (
+        "%(asctime)s [%(name)s:%(filename)s:%(lineno)d] %(levelname)s: %(message)s"
+    )
+    log_config_dict = {
+        "version": 1,
+        "disable_existing_loggers": False,
+        "loggers": {
+            "root": {
+                "level": log_level,
+                "handlers": ["console"],
+            },
+            "bugzooka": {
+                "level": log_level,
+                "handlers": ["console"],
+                "propagate": False,
+            },
+        },
+        "handlers": {
+            "console": {
+                "class": "logging.StreamHandler",
+                "formatter": "standard",
+            },
+        },
+        "formatters": {
+            "standard": {"format": log_msg_fmt},
+        },
+    }
+
+    logging.config.dictConfig(log_config_dict)
