+++ bugzooka/integrations/slack_fetcher.py
@@ -1,6 +1,3 @@
-import argparse
-import logging
-import os
 import io
 import signal
 import sys
@@ -9,24 +6,23 @@ import time
 from slack_sdk import WebClient
 from slack_sdk.errors import SlackApiError
 
-from src.config import SLACK_BOT_TOKEN, SLACK_CHANNEL_ID, JEDI_BOT_SLACK_USER_ID, get_product_config
-from src.constants import (
+from bugzooka.core.config import (
+    SLACK_BOT_TOKEN,
+    JEDI_BOT_SLACK_USER_ID,
+)
+from bugzooka.core.constants import (
     MAX_CONTEXT_SIZE,
     MAX_PREVIEW_CONTENT,
-    SLACK_POLL_INTERVAL,
 )
-from src.logging import configure_logging
-from src.log_analyzer import (
+from bugzooka.analysis.log_analyzer import (
     download_and_analyze_logs,
     filter_errors_with_llm,
     run_agent_analysis,
 )
-from src.inference import (
+from bugzooka.integrations.inference import (
     InferenceAPIUnavailableError,
     AgentAnalysisLimitExceededError,
 )
-
-from src.utils import get_slack_message_blocks
 
 
 class SlackMessageFetcher:
@@ -50,6 +46,42 @@ class SlackMessageFetcher:
         # Handle SIGINT (Ctrl+C) for graceful exit
         signal.signal(signal.SIGINT, self.shutdown)
 
+    def _get_slack_message_blocks(
+        self, markdown_header, content_text, use_markdown=False
+    ):
+        """
+        Prepares a slack message building blocks
+
+        :param markdown_header: markdown header to be displayed
+        :param content_text: text message content (preformatted or markdown)
+        :param use_markdown: if True, render content as markdown; if False, use preformatted text
+        :return: a sanitized version of text blocks
+        """
+        header_block = {
+            "type": "section",
+            "text": {"type": "mrkdwn", "text": markdown_header},
+        }
+
+        if use_markdown:
+            content_block = {
+                "type": "markdown",
+                "text": content_text.strip(),
+            }
+        else:
+            content_block = {
+                "type": "rich_text",
+                "block_id": "error_logs_block",
+                "elements": [
+                    {
+                        "type": "rich_text_preformatted",
+                        "elements": [{"type": "text", "text": content_text.strip()}],
+                        "border": 0,
+                    }
+                ],
+            }
+
+        return [header_block, content_block]
+
     def _filter_new_messages(self, messages):
         """Filter messages to only include new ones that haven't been processed."""
         new_messages = []
@@ -60,20 +92,27 @@ class SlackMessageFetcher:
             replies = self.client.conversations_replies(channel=self.channel_id, ts=ts)
             messages_in_thread = replies.get("messages", [])
             bot_replied = any(
-                reply.get("user") == JEDI_BOT_SLACK_USER_ID for reply in messages_in_thread[1:]  # skip the parent msg at index 0
+                reply.get("user") == JEDI_BOT_SLACK_USER_ID
+                for reply in messages_in_thread[1:]  # skip the parent msg at index 0
             )
-            if self.last_seen_timestamp is None or float(ts) > float(self.last_seen_timestamp):
+            if self.last_seen_timestamp is None or float(ts) > float(
+                self.last_seen_timestamp
+            ):
                 if bot_replied:
-                    self.logger.debug(f"Skipping message with timestamp {ts} due to bot replied")
+                    self.logger.debug(
+                        f"Skipping message with timestamp {ts} due to bot replied"
+                    )
                 else:
                     new_messages.append(msg)
             else:
                 self.logger.debug(
                     f"Skipping message with timestamp {ts} due to timestamp filter"
-                    )
+                )
         return new_messages
 
-    def _send_error_logs_preview(self, errors_list, categorization_message, max_ts, is_install_issue=False):
+    def _send_error_logs_preview(
+        self, errors_list, categorization_message, max_ts, is_install_issue=False
+    ):
         """Send error logs preview to Slack (either as message or file)."""
         errors_log_preview = "\n".join(errors_list or [])[:MAX_PREVIEW_CONTENT]
         errors_list_string = "\n".join(errors_list or [])[:MAX_CONTEXT_SIZE]
@@ -97,7 +136,7 @@ class SlackMessageFetcher:
             )
         else:
             self.logger.info("📤 Trying to just send the preview message")
-            message_block = get_slack_message_blocks(
+            message_block = self._get_slack_message_blocks(
                 markdown_header=f":checking: *Error Logs Preview ({categorization_message})*\n",
                 content_text=f"{errors_log_preview.strip()}",
             )
@@ -113,7 +152,7 @@ class SlackMessageFetcher:
                 "This appears to be an installation or maintenance issue. "
                 "Please re-trigger the run."
             )
-            message_block = get_slack_message_blocks(
+            message_block = self._get_slack_message_blocks(
                 markdown_header=":repeat: *Re-trigger Suggested*\n",
                 content_text=retrigger_message,
             )
@@ -126,7 +165,7 @@ class SlackMessageFetcher:
 
     def _send_analysis_result(self, response, max_ts):
         """Send the final analysis result to Slack."""
-        message_block = get_slack_message_blocks(
+        message_block = self._get_slack_message_blocks(
             markdown_header=":fast_forward: *Implications to understand (AI generated) *\n",
             content_text=response,
             use_markdown=True,
@@ -145,7 +184,7 @@ class SlackMessageFetcher:
             "The inference API is currently unavailable. "
             "Raw error logs have been provided above for manual review."
         )
-        message_block = get_slack_message_blocks(
+        message_block = self._get_slack_message_blocks(
             markdown_header=":warning: *Analysis Unavailable*\n",
             content_text=fallback_message,
         )
@@ -156,7 +195,9 @@ class SlackMessageFetcher:
             thread_ts=max_ts,
         )
 
-    def _process_message(self, msg, product, ci_system, product_config, enable_inference):
+    def _process_message(
+        self, msg, product, ci_system, product_config, enable_inference
+    ):
         """Process a single message through the complete pipeline."""
         user = msg.get("user", "Unknown")
         text = msg.get("text", "No text available")
@@ -169,12 +210,16 @@ class SlackMessageFetcher:
             return ts
 
         # Extract and download logs
-        errors_list, categorization_message, requires_llm, is_install_issue = download_and_analyze_logs(text, ci_system)
+        errors_list, categorization_message, requires_llm, is_install_issue = (
+            download_and_analyze_logs(text, ci_system)
+        )
         if errors_list is None:
             return ts
 
         # Send error logs preview first
-        self._send_error_logs_preview(errors_list, categorization_message, ts, is_install_issue)
+        self._send_error_logs_preview(
+            errors_list, categorization_message, ts, is_install_issue
+        )
 
         if is_install_issue or not enable_inference:
             return ts
@@ -245,7 +290,9 @@ class SlackMessageFetcher:
                         msg, product, ci_system, product_config, enable_inference
                     )
 
-                    if processed_ts and float(processed_ts) > float(self.last_seen_timestamp or 0):
+                    if processed_ts and float(processed_ts) > float(
+                        self.last_seen_timestamp or 0
+                    ):
                         self.logger.info(
                             f"Updating last_seen_timestamp from {self.last_seen_timestamp} to {processed_ts}"
                         )
@@ -288,62 +335,3 @@ class SlackMessageFetcher:
         self.logger.info("🛑 Received exit signal. Stopping message fetcher...")
         self.running = False
         sys.exit(0)
-
-
-def str_to_bool(value):
-    """Convert string ENV value to bool."""
-    return str(value).lower() == "true"
-
-
-# export PYTHONPATH=$(pwd)/src:$PYTHONPATH
-if __name__ == "__main__":
-    VALID_LOG_LEVELS = ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
-    parser = argparse.ArgumentParser(description="Slack Log Analyzer Bot")
-
-    parser.add_argument(
-        "--product",
-        type=str,
-        default=os.environ.get("PRODUCT"),
-        help="Product type (e.g., openshift, ansible)",
-    )
-    parser.add_argument(
-        "--ci", type=str, default=os.environ.get("CI"), help="CI system name"
-    )
-    parser.add_argument(
-        "--log-level",
-        type=str,
-        choices=VALID_LOG_LEVELS,
-        default=os.environ.get("LOG_LEVEL", "INFO").upper(),
-        help="Logging level (e.g., DEBUG, INFO, WARNING, ERROR, CRITICAL)."
-        " Can also be set via LOG_LEVEL env var",
-    )
-    parser.add_argument(
-        "--enable-inference",
-        action="store_true",
-        default=str_to_bool(os.environ.get("ENABLE_INFERENCE", "false")),
-        help="Enable inference mode. Can also be set via ENABLE_INFERENCE env var (true/false).",
-    )
-
-    args = parser.parse_args()
-    configure_logging(args.log_level)
-    logger = logging.getLogger(__name__)
-    missing_args = []
-    if not args.product:
-        missing_args.append("product or PRODUCT")
-    if not args.ci:
-        missing_args.append("ci or CI")
-    if missing_args:
-        logger.error("Missing required arguments or env vars: {%s}", missing_args)
-        sys.exit(1)
-
-    kwargs = {
-        "product": args.product.upper(),
-        "ci": args.ci.upper(),
-        "product_config": get_product_config(product_name=args.product.upper()),
-        "enable_inference": args.enable_inference,
-    }
-
-    fetcher = SlackMessageFetcher(
-        channel_id=SLACK_CHANNEL_ID, logger=logger, poll_interval=SLACK_POLL_INTERVAL
-    )
-    fetcher.run(**kwargs)
