+++ requirements.txt
@@ -48,7 +48,7 @@ jsonpointer==3.0.0
 keyring==25.6.0
 langchain==0.3.26
 langchain-community==0.3.27
-langchain-core==0.3.67
+langchain-core>=0.3.68,<0.4.0
 langchain-text-splitters==0.3.8
 langsmith==0.4.4
 logmine==0.4.1
